chore(chromatic): reduce chromatic workflow runs (#155)
We're close to our snapshot limit as a Free org in chromatic (4500 out of 5000). Chromatic provided default config runs on every push, but our workflow only requires it to run on reviewable PR's and their updates. I implemented an attempt at the goal, which I researched [here](https://github.community/t/dont-run-actions-on-draft-pull-requests/16817).